用户注册自动创建订单

+++ src/main/java/com/cdp/zwy/buildbody/module/system/service/impl/SysUserServiceImpl.java
@@ -170,7 +170,7 @@ public class SysUserServiceImpl extends ServiceImpl<SysUserDao, SysUser> impleme
             cardTimes = cardTemplate.getTimes();
         }
         
-        Long orderId = sysOrderService.createMemberCardOrder(user.getUserId(), cardTimes, cardTemplate.getPrice());
+        Long orderId = sysOrderService.createMemberCardOrder(user.getUserId(), cardTimes, cardTemplate.getPrice().doubleValue());
         // 自动支付订单（实际项目中应该有支付流程）
         sysOrderService.payOrder(orderId);
